refactor: productItem 파일분리 적용
